Collaboration and data section fixes

+++ packages/builder/src/stores/backend/datasources.js
@@ -1,8 +1,13 @@
 import { writable, derived, get } from "svelte/store"
-import { IntegrationTypes, DEFAULT_BB_DATASOURCE_ID } from "constants/backend"
-import { queries, tables } from "./"
+import {
+  IntegrationTypes,
+  DEFAULT_BB_DATASOURCE_ID,
+  BUDIBASE_INTERNAL_DB_ID,
+} from "constants/backend"
+import { tables, queries } from "./"
 import { API } from "api"
 import { DatasourceFeature } from "@budibase/types"
+import { TableNames } from "constants"
 
 export class ImportTableError extends Error {
   constructor(message) {
@@ -23,13 +28,40 @@ export function createDatasourcesStore() {
     schemaError: null,
   })
 
-  const derivedStore = derived(store, $store => ({
-    ...$store,
-    selected: $store.list?.find(ds => ds._id === $store.selectedDatasourceId),
-    hasDefaultData: $store.list.some(
-      datasource => datasource._id === DEFAULT_BB_DATASOURCE_ID
-    ),
-  }))
+  const derivedStore = derived([store, tables], ([$store, $tables]) => {
+    // Set the internal datasource entities from the table list, which we're
+    // able to keep updated unlike the egress generated definition of the
+    // internal datasource
+    let internalDS = $store.list?.find(ds => ds._id === BUDIBASE_INTERNAL_DB_ID)
+    let otherDS = $store.list?.filter(ds => ds._id !== BUDIBASE_INTERNAL_DB_ID)
+    if (internalDS) {
+      internalDS = {
+        ...internalDS,
+        entities: $tables.list?.filter(table => {
+          return (
+            table.sourceId === BUDIBASE_INTERNAL_DB_ID &&
+            table._id !== TableNames.USERS
+          )
+        }),
+      }
+    }
+
+    // Build up enriched DS list
+    // Only add the internal DS if we have at least one non-users table
+    let list = []
+    if (internalDS?.entities?.length) {
+      list.push(internalDS)
+    }
+    list = list.concat(otherDS || [])
+
+    return {
+      ...$store,
+      list,
+      selected: list?.find(ds => ds._id === $store.selectedDatasourceId),
+      hasDefaultData: list?.some(ds => ds._id === DEFAULT_BB_DATASOURCE_ID),
+      hasData: list?.length > 0,
+    }
+  })
 
   const fetch = async () => {
     const datasources = await API.getDatasources()
@@ -50,20 +82,14 @@ export function createDatasourcesStore() {
 
   const updateDatasource = response => {
     const { datasource, error } = response
-    store.update(state => {
-      const currentIdx = state.list.findIndex(ds => ds._id === datasource._id)
-      const sources = state.list
-      if (currentIdx >= 0) {
-        sources.splice(currentIdx, 1, datasource)
-      } else {
-        sources.push(datasource)
-      }
-      return {
-        list: sources,
-        selectedDatasourceId: datasource._id,
+    if (error) {
+      store.update(state => ({
+        ...state,
         schemaError: error,
-      }
-    })
+      }))
+    }
+    replaceDatasource(datasource._id, datasource)
+    select(datasource._id)
     return datasource
   }
 
@@ -134,18 +160,14 @@ export function createDatasourcesStore() {
   }
 
   const deleteDatasource = async datasource => {
+    if (!datasource?._id || !datasource?._rev) {
+      return
+    }
     await API.deleteDatasource({
-      datasourceId: datasource?._id,
-      datasourceRev: datasource?._rev,
+      datasourceId: datasource._id,
+      datasourceRev: datasource._rev,
     })
-    store.update(state => {
-      const sources = state.list.filter(
-        existing => existing._id !== datasource._id
-      )
-      return { list: sources, selected: null }
-    })
-    await queries.fetch()
-    await tables.fetch()
+    replaceDatasource(datasource._id, null)
   }
 
   const removeSchemaError = () => {
@@ -154,7 +176,6 @@ export function createDatasourcesStore() {
     })
   }
 
-  // Handles external updates of datasources
   const replaceDatasource = (datasourceId, datasource) => {
     if (!datasourceId) {
       return
@@ -166,6 +187,8 @@ export function createDatasourcesStore() {
         ...state,
         list: state.list.filter(x => x._id !== datasourceId),
       }))
+      tables.removeDatasourceTables(datasourceId)
+      queries.removeDatasourceQueries(datasourceId)
       return
     }
